commit_64572 ------------------------------------------------------------------------ r64572 | vboxsync | 2016-11-04 20:54:38 +0800 (Fri, 04 Nov 2016) | 3 lines
Audio/HDA: Rewritten large parts of the interrupt delivery and DMA handling to improve audio data delivery.

We now update the interrupt status when needed, and not when the status is read by the guest -- this created all kinds of funny side effects. Also, split out the DMA transfers to an own routine for easier debugging and made sure we transfer the correct sizes according to FIFOS boundaries.
------------------------------------------------------------------------