commit_64788 ------------------------------------------------------------------------ r64788 | vboxsync | 2016-12-06 19:39:49 +0800 (Tue, 06 Dec 2016) | 1 line
bugref:8679: Additions/x11/graphics: ldconfig on Debian 7.11 guest not picking up VBoxOGL.so: old versions of Mesa contain a GNU ABI notes section which makes ldconfig always prefer them to available alternatives, regardless of the specified order of precedence.  This was fixed for most current distributions in freedesktop.org bug 26663.  Handle this un-invasively by detecting the problem and telling the user how to fix it.
------------------------------------------------------------------------