commit_64830 ------------------------------------------------------------------------ r64830 | vboxsync | 2016-12-12 17:55:37 +0800 (Mon, 12 Dec 2016) | 1 line
bugref:8677: Additions/Graphics/Wayland: Fedora 25 issues: remove functionality from the vboxvideo drm driver for moving unused buffers in video RAM to system RAM.  This was copied and pasted from the driver for the memory-strapped AST driver, and not so relevant for us.  However it was having the unintended consequence that when GNOME Shell page-flipped in a new buffer it was ending up at the same video RAM address as the old one (I unfortunately currently do not understand the details), leading the graphics card to think that nothing had changed on-screen.  Now unaccelerated updates are working correctly in multi-monitor Fedora 25 guests, though probably still not very efficiently.
------------------------------------------------------------------------