commit_64894 ------------------------------------------------------------------------ r64894 | vboxsync | 2016-12-16 08:43:05 +0800 (Fri, 16 Dec 2016) | 1 line
asn1-ut-core.cpp: Encoding fix for Weird vtable for lone RTASN1CORE structures to prevent them from beling left out of the output.  Because we have an pfnEnum callback and doesn't actually do anything, we need a pfnEncodePrep callback that doesn't do anything either and a pfnEncodeWrite that does what it says on the label.
------------------------------------------------------------------------