commit_65335 ------------------------------------------------------------------------ r65335 | vboxsync | 2017-01-16 22:01:48 +0800 (Mon, 16 Jan 2017) | 1 line
testmanager,testdriver: Adjustments of process info dumping. Don't do it if we're low on memory (gdb uses lots of memory!). Changed the kind from 'log/host/vmprocess' to 'process/report/vm' to better fit with the rest. processGetInfo needs a parameter indicating whether or not to try sudo the operations.
------------------------------------------------------------------------