commit_65516 ------------------------------------------------------------------------ r65516 | vboxsync | 2017-01-30 17:55:26 +0800 (Mon, 30 Jan 2017) | 2 lines
DevPciIch9: if we encode 64-bit prefetchable addresses, set the lower 4 bits of the PREF_BASE and PREF_LIMIT registers to 1 (according to the PCI-to-PCI bridge spec). In the 'info pci' command properly print "no I/O", "no memory" or "no prefetch memory" if the limit register is programmed to a smaller value than the base register.

------------------------------------------------------------------------