commit_65544 ------------------------------------------------------------------------ r65544 | vboxsync | 2017-01-31 22:58:40 +0800 (Tue, 31 Jan 2017) | 1 line
Main/HostUSBDeviceImpl: Report the port number on Windows for all USB devices. In case of multiple active backends active (USB/IP) which may contain a valid port number those would be reported as 0 too (breaks USB testing if the capture option is enabled). Forcing the port number to be 0 should be done in the Windows specific part of the USB code if it turns out that it is still necessary for devices attached to a Windows host
------------------------------------------------------------------------