commit_65563 ------------------------------------------------------------------------ r65563 | vboxsync | 2017-02-01 19:51:44 +0800 (Wed, 01 Feb 2017) | 3 lines
bugref:8748: Additions/Graphics/Wayland: investigate EGLStreams support feasibility: modify guest GLX visual and FBConfig code to be independent of guest X11 visual configuration.  Now we just report one of each with 32-bit ARGB instead of one for every X11 visual on the guest, some with 0-bit alpha and some with 8-bit.  There was a comment in the code that we needed both 24-bit and 32-bit.  I do not know why we should need 24-bit.

Change tested with a couple Unity and GNOME Shell versions, and glxinfo, glxgears, eglinfo and eglgears.  More good test cases would be good.  glxmark2 fails with both the old and the new code.
------------------------------------------------------------------------