commit_65729 ------------------------------------------------------------------------ r65729 | vboxsync | 2017-02-10 21:07:38 +0800 (Fri, 10 Feb 2017) | 2 lines
ValidationKit: when setting the MAC address: if we can't find out the IP address of the system, it's better to make the method fail, because that makes the VM creation and thus the test fail properly. The previous way of continuing silently (but with an increased error counter) leads to very delayed test failures which can't be easily identified.

------------------------------------------------------------------------