commit_65808 ------------------------------------------------------------------------ r65808 | vboxsync | 2017-02-20 19:41:59 +0800 (Mon, 20 Feb 2017) | 2 lines
Devices/Network/lwip-new: the mailbox uses pointers to the message. As we don't know when the message is consumed, don't pass a pointer to a local stack variable but always allocate a message. This could fix a potential crash of VBoxNetNAT during shutdown (when handling the a problem in our code for terminating the lwip stack. A mailbox is used to store the pointer to the TCPIP_MSG_CALLBACK_TERMINATE message).

------------------------------------------------------------------------