commit_65957 ------------------------------------------------------------------------ r65957 | vboxsync | 2017-03-07 04:41:43 +0800 (Tue, 07 Mar 2017) | 1 line
PGMPool: Address problem where we would incorrectly try to flush an active CR3 page (or pages in PAE) because we thought it was being reused, which obviously would fail, then pretend we succeeded and restart the instruction.  Happend on a STOS during linux bootup (e.g. ubuntu 15.10/amd64).  Fix was to extend pgmRZPoolMonitorIsReused with a check on locked pages.
------------------------------------------------------------------------