commit_66544 ------------------------------------------------------------------------ r66544 | vboxsync | 2017-04-13 01:02:30 +0800 (Thu, 13 Apr 2017) | 2 lines
bugref:8524: Additions/linux: play nicely with distribution-installed Additions
Change header of files which are expected to end up in the Linux kernel to the MIT licence to simplify life for people wanting to port vboxvideo to other kernels and to simplify synchronising changes back to VirtualBox.  Update author information in files which have it, but do not add it to files which do not.
------------------------------------------------------------------------