commit_66679 ------------------------------------------------------------------------ r66679 | vboxsync | 2017-04-26 23:24:20 +0800 (Wed, 26 Apr 2017) | 14 lines
bugref:8524: Additions/linux: play nicely with distribution-installed Additions
[PATCH] Additions/common/VBoxVideo/HGSMIBase: Group setup functions
 with related functions

Group the VBoxHGSMISetupHostContext function together with the
functions using PHGSMIHOSTCOMMANDCONTEXT and group
VBoxHGSMISetupGuestContext together with the functions using
PHGSMIGUESTCOMMANDCONTEXT. This makes it easier to later #ifdef
out the bits we are not going to use in the Linux vboxvideo drm driver.

Note this commit makes no functional changes it just moves a few
blocks of code around.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
------------------------------------------------------------------------