commit_66940 ------------------------------------------------------------------------ r66940 | vboxsync | 2017-05-18 00:44:04 +0800 (Thu, 18 May 2017) | 2 lines
Main/NetworkAdapter+NATEngine: Improve default handling (deal with NULL guest OS type), and also add a check if the current settings are default. Note that the API defaults are not the same as the Settings defaults, and that's intentional as the API defaults apply to NICs which the settings didn't read from the config at all.

------------------------------------------------------------------------