commit_67043 ------------------------------------------------------------------------ r67043 | vboxsync | 2017-05-23 22:07:04 +0800 (Tue, 23 May 2017) | 2 lines
bugref:8533: Additions/x11: fully support VMSVGA
Add VBoxClient support for dynamic resizing with VMSVGA enabled.  This is not yet started by default.  VBoxClient --vmsvga should be run as root outside of an X session for recent versions of the Linux kernel and interacts with the SVGA kernel graphics driver.  When run inside an X session it interacts with the X11 graphics driver.
------------------------------------------------------------------------