commit_67053 ------------------------------------------------------------------------ r67053 | vboxsync | 2017-05-24 16:02:36 +0800 (Wed, 24 May 2017) | 2 lines
bugref:8834: Makeself installers/Linux: indicate problems better in the exit codes
Do not attempt to start the services part of VirtualBox on Linux hosts if building the kernel modules fails.  This makes it easier to test an installer failure path, as starting the services caused another build attempt to be started with default settings, which made it hard to cause a build failure.
------------------------------------------------------------------------