commit_67139 ------------------------------------------------------------------------ r67139 | vboxsync | 2017-05-30 19:17:28 +0800 (Tue, 30 May 2017) | 1 line
ValidationKit/TXS: Added a configurable wait on the CD-ROM during auto upgrade to work around a solaris problem where it takes a little while before it is mounted.  The wait defaults to 8 seconds on solaris and 1 second on the other OSes.  Ideally, we should figure out that a medium is present before wasting time waiting, but that's probably complicated and can wait.  The wait can be configured using the --wait-cdrom <secs> option.  Setting it to zero disables the wait.  There is no upper limit.
------------------------------------------------------------------------