commit_67151 ------------------------------------------------------------------------ r67151 | vboxsync | 2017-05-31 15:12:18 +0800 (Wed, 31 May 2017) | 2 lines
Runtime/r3/socket: when checking for select() results it is possible that pThis->hNative became invalid in the meantimebecause the socket was closed, see rtSocketCloseIt() where we invalid pThis->hNative first and then close the socket. We could also remember hNative for the correct test after select() but I don't think we really care about the RTSOCKET_EVT_READ result on closed sockets.

------------------------------------------------------------------------