commit_67356 ------------------------------------------------------------------------ r67356 | vboxsync | 2017-06-13 19:22:15 +0800 (Tue, 13 Jun 2017) | 2 lines
EFI/Firmware: Eliminate the use of buffers in the blitting code (especially the full frame buffer for video to video blitting was really hurting as for high resolutions it needed a lot of RAM), primarily by cleaning up the code to be for 32bpp only, which it already was previously. Fixed UGA mode initialization, as it only pretended to set up the specified mode, but actually was never using anything else but mode 0 due to logical flaws.

------------------------------------------------------------------------