commit_67587 ------------------------------------------------------------------------ r67587 | vboxsync | 2017-06-24 00:18:49 +0800 (Sat, 24 Jun 2017) | 2 lines
DevVGA: Add a few missing checks which skip the VRAM alias mapping to the LFB when the PCI BAR isn't set up (i.e. the LFB physical address is 0, conflicting with system RAM). Performance is very bad in this case, but it's not an expected use case to have incorrectly initialized BARs.

------------------------------------------------------------------------