commit_67792 ------------------------------------------------------------------------ r67792 | vboxsync | 2017-07-05 19:30:48 +0800 (Wed, 05 Jul 2017) | 2 lines
Main/NetworkAdapterImpl: fixed another case with wrong NAT/DHCP refcounting: don't change the NAT refcount if the current networking mode is not 'NAT network'. This happens for instance if we change the adapter type from something else to 'NAT network' in the GUI and at the same time change the NAT network name.

------------------------------------------------------------------------