commit_67852 ------------------------------------------------------------------------ r67852 | vboxsync | 2017-07-07 21:34:33 +0800 (Fri, 07 Jul 2017) | 7 lines
bugref:8906: Additions/linux: kernel modules not always rebuild before reboot on Fedora kernel upgrade
Make sure vboxvideo is loaded when it is rebuilt at boot time.

Our Additions boot scripts load vboxvideo manually if it was previously
blacklisted.  However, there is another case when it might not be loaded
automatically: when it has just been built at boot time.  So load it
unconditionally if it is not blacklisted.
------------------------------------------------------------------------