commit_67858 ------------------------------------------------------------------------ r67858 | vboxsync | 2017-07-07 23:24:25 +0800 (Fri, 07 Jul 2017) | 7 lines
bugref:8906: Additions/linux: kernel modules not always rebuild before reboot on Fedora kernel upgrade
Load vboxvideo when it is rebuilt at boot time even on guests without Xorg.

Our Additions boot scripts load vboxvideo manually if it was previously
blacklisted.  However, there is another case when it might not be loaded
automatically: when it has just been built at boot time and Xorg is not.
installed on the guest.  Fix that.
------------------------------------------------------------------------