commit_68027 ------------------------------------------------------------------------ r68027 | vboxsync | 2017-07-18 22:21:17 +0800 (Tue, 18 Jul 2017) | 8 lines
bugref:3809: Linux installer maintenance
Clean up unneeded kernel modules when the vboxdrv service is started.

When a Linux kernel is removed on the host for which VirtualBox driver
modules were installed, the kernel directory and those modules were left
in place.  This change adds a check to the vboxdrv service script to look
for such orphaned modules on start-up and remove them along with the
directories.
------------------------------------------------------------------------