commit_68048 ------------------------------------------------------------------------ r68048 | vboxsync | 2017-07-19 23:02:01 +0800 (Wed, 19 Jul 2017) | 2 lines
Main/Settings.cpp: Use a "Description" XML tag instead of an attribute for the Medium registry entries. Matches what the code should've been doing since day 1 (and the XML schema), but unfortunately since VirtualBox 3.2 the code used the attribute. Not that there was any way to set the description in the API before 5.2.

------------------------------------------------------------------------