commit_68089 ------------------------------------------------------------------------ r68089 | vboxsync | 2017-07-23 18:07:12 +0800 (Sun, 23 Jul 2017) | 11 lines
bugref:8778: Double click not working with precision touchpad (user report)
A new attempt to fix this problem by extending the throttle timer test for
unreported events to also check whether the accumulated button field is
non-zero.

Also add a test case which is executed on device construction in strict
builds to check that the use case to be fixed works and that the mouse
dragging regression introduced by the first fix attempt is not present.
Tested the test case on the two previous fix attempts which were known to
respectively fix one case and break the other.  The current fix attempt
passes both tests.
------------------------------------------------------------------------