commit_68604 ------------------------------------------------------------------------ r68604 | vboxsync | 2017-09-01 23:32:33 +0800 (Fri, 01 Sep 2017) | 1 line
bs3kit.h: Made Bs3StrFormatV work in small model 16-bit environment too.  va_list is tricky in this context, because it's an C array and gets treated like a pointer when passed as an argument.  so, it ended up being a near pointer even when -zu (SS!=DS) was enabled and crashed.  Fix was to add BS3_FAR to the va_list arguments.  Also, undefine IN_RING0 before forcibly defining it, since it is 1 when defined on the command line (e.g. for VBoxMouse) and causes a warning/error.
------------------------------------------------------------------------