commit_68699 ------------------------------------------------------------------------ r68699 | vboxsync | 2017-09-07 23:12:54 +0800 (Thu, 07 Sep 2017) | 1 line
pdmifs.h,Serial: Reworked stream interface. The old design with the two read/write threads had a race where the read thread could access already destroyed VMM structures during destruction if data was read. This was solved by adding a poll callback which waits for data to arrive and which can be interrupt to make the thread respond to VM state changes and suspend before destruction starts. This required reworking all the drivers using it. DrvTCP was reworked to make use of the RTTcp*, RTSocket* and RTPoll* API in that process to get rid of platform dependent code there (which wasn't all available when the driver was createt).
------------------------------------------------------------------------