commit_69428 ------------------------------------------------------------------------ r69428 | vboxsync | 2017-10-27 21:06:27 +0800 (Fri, 27 Oct 2017) | 8 lines
Installer/linux/vboxdrv.sh: always clean up empty kernel driver misc folders.
bugref:3809: Linux installer maintenance
When we clean up and remove our host kernel drivers, we also check to see if
we are leaving behind an empty misc folder in the kernel driver tree, and if
so we delete it.  However we were not always as careful in the past, with
the result that many systems have a number of driver folders for different
kernel versions, empty except for an empty misc folder.  This change makes
us delete those if we find them during clean-up.
------------------------------------------------------------------------