commit_69529 ------------------------------------------------------------------------ r69529 | vboxsync | 2017-10-31 01:42:36 +0800 (Tue, 31 Oct 2017) | 7 lines
Linux drivers: make it easier to use different kernel headers.
bugref:4567: Linux kernel driver maintenance

With this change it is possible to pass KERN_DIR on the command line to
specify which kernel headers to do test builds of kernel modules against.
The default is now the most recent headers found, not the matching headers
for the currently running kernel.
------------------------------------------------------------------------