commit_69563 ------------------------------------------------------------------------ r69563 | vboxsync | 2017-11-03 22:32:12 +0800 (Fri, 03 Nov 2017) | 7 lines
Additions/linux/drm: better detection for EL7 variants.
bugref:4567: Linux kernel driver maintenance

r118292/r68647 was very invasive and broke Debian 9 on top.  This change
refines the check for supported guests for the drm driver by checking the
kernel version string to be sure that it either starts with a supported
version (that is, not 3.10.x or older) or contains the EL7 signature.
------------------------------------------------------------------------