commit_69892 ------------------------------------------------------------------------ r69892 | vboxsync | 2017-12-01 06:34:01 +0800 (Fri, 01 Dec 2017) | 1 line
Runtime: Start on a platform independent serial port access API which will be more advanced and cleaner than the code in DrvHostSerial. DrvHostSerial will be converted to use this API later on when it is mature. Additionally it will be used by a serial port test utility for automatic serial port testing in the validation kit
------------------------------------------------------------------------