commit_70050 ------------------------------------------------------------------------ r70050 | vboxsync | 2017-12-11 01:28:20 +0800 (Mon, 11 Dec 2017) | 1 line
Devices/OHCI: Don't touch others member of the endpoint descriptor when writing the head pointer HeadP and embedded flags back to the guest as it might corrupt data updated by the guest inbetween. This can result in TDs not being recognized later on because the linked list contains old data again. Some cleanups for the caching to make it more contained and less spread out over the code
------------------------------------------------------------------------