commit_70141 ------------------------------------------------------------------------ r70141 | vboxsync | 2017-12-15 20:25:49 +0800 (Fri, 15 Dec 2017) | 9 lines
http-curl.cpp: rtHttpWinConfigureProxyForUrl - also assume no proxy if
we get ERROR_WINHTTP_UNABLE_TO_DOWNLOAD_SCRIPT.  This avoid assertion
in the debug build.

That error is what I get on my home network without automatic proxy,
not ERROR_WINHTTP_AUTODETECTION_FAILED.  The latter is not listed as a
possible error for WinHttpGetProxyForUrl function(), but for
WinHttpDetectAutoProxyConfigUrl().

------------------------------------------------------------------------