commit_70145 ------------------------------------------------------------------------ r70145 | vboxsync | 2017-12-15 20:52:29 +0800 (Fri, 15 Dec 2017) | 4 lines
VMM/IEM: Update the address size field for basic IN/OUT instr. and don't leave it 0. Real hardware doesn't, and
I think using the effective address size is the right thing to do here given the instructions only have an immediate
or register operand and question of address size is kind of moot, so just follow real hardware here.

------------------------------------------------------------------------