commit_70407 ------------------------------------------------------------------------ r70407 | vboxsync | 2018-01-02 04:38:02 +0800 (Tue, 02 Jan 2018) | 1 line
Additions,ValidationKit,VBoxPeSetVersion: Force linker to split .bss out from the .data section as NT 3.1 cannot deal with that (it doesn't check the Misc.VirtualSize field and maps raw bytes up to the next section).  We tell the linker .bss is uncached, and have VBoxPeSetImage undo this change. [build fix]
------------------------------------------------------------------------