commit_70726 ------------------------------------------------------------------------ r70726 | vboxsync | 2018-01-24 21:58:10 +0800 (Wed, 24 Jan 2018) | 7 lines
Do not allow SUID, SGID or sticky bits on shared folders.
bugref:9047: Clean up Linux guest vboxuser device
Shared folders are primarily intended as a convenient method of file transfer
between guest and host, but not for more complicated things.  Therefore the
SUID, SGID or sticky bits do not make sense there, and this change removes
them so that people do not get wrong ideas.
Not strictly related to the vboxuser device, but same general idea.
------------------------------------------------------------------------