commit_71155 ------------------------------------------------------------------------ r71155 | vboxsync | 2018-02-28 22:56:07 +0800 (Wed, 28 Feb 2018) | 1 line
Devices/Bus/ICH9: Don't meddle with the IRQ value when traversing bridges if the originating device has MSI/MSI-X enabled. The IRQ has a completely different meaning here (the MSI/MSI-X vector to use) and is not subject to the pin routing for bridges. MSI/MSI-X interrupt handling code paths should really be separated properly to avoid such confusing issues but for a safe backport this is the immediate fix for now
------------------------------------------------------------------------