commit_71886 ------------------------------------------------------------------------ r71886 | vboxsync | 2018-04-18 04:00:55 +0800 (Wed, 18 Apr 2018) | 6 lines
Linux Additions/3D: add symbols needed by Mutter in Ubuntu 18.04.
#9162: Fix linking trouble on Ubuntu 18.04 with our GA-provided fake libGL.so
Add some symbols needed by Mutter on Ubuntu 18.04 in a rather hacky way.
All of them are simply renamed versions of symbols that we already provide.
It is not unlikely that we will need to add more in the same way in future.
Fix a Windows and OS X burn.
------------------------------------------------------------------------