commit_72091 ------------------------------------------------------------------------ r72091 | vboxsync | 2018-05-03 20:53:03 +0800 (Thu, 03 May 2018) | 6 lines
Address sanitiser: split up option into statically and dynamically linked.
bugref:8019: Undefined behaviour sanitiser
When we link the address sanitiser libraries dynamically we need to set
LD_LIBRARY_PATH for the 3D test to work.  However, when we link statically
the test suite causes problems.  So have an option for both and let the
developer choose.
------------------------------------------------------------------------