commit_72093 ------------------------------------------------------------------------ r72093 | vboxsync | 2018-05-03 22:02:30 +0800 (Thu, 03 May 2018) | 5 lines
Restrict old glibc support to Guest Additions only.
bugref:9112: Linux Additions: fix building binaries for older Linux user space
Old glibc support is only of interest when building the Guest Additions, so
this change restricts it to binaries which will become part of the Additions,
so that it can be enabled by default.
------------------------------------------------------------------------