commit_72137 ------------------------------------------------------------------------ r72137 | vboxsync | 2018-05-07 20:31:41 +0800 (Mon, 07 May 2018) | 7 lines
Some more fixes for building backwards-compatible Linux Additions binaries.
bugref:9112: Linux Additions: fix building binaries for older Linux user space
* Only link with --hash-style=sysv where that option is supported.
* Fix chk_stubs.c build mess by building it as a normal source file.  It does
not need to come after the static libraries using its symbols on the command
line as it is an object file, not a library, and therefore its symbols are
always pulled in.
------------------------------------------------------------------------