commit_72571 ------------------------------------------------------------------------ r72571 | vboxsync | 2018-06-16 04:36:17 +0800 (Sat, 16 Jun 2018) | 1 line
Runtime/RTFuzz: Started to extend the master command to provide a network interface for controlling the fuzzing process. The interface uses JSON requests and responses to control the fuzzing process, like creating a new fuzzing run and suspending, resuming and stopping it. The JSON request contains everything required (except for the binary itself) for easier control of multiple masters on different hosts via a central control mechanism (later integration into the validation kit).
------------------------------------------------------------------------