commit_72627 ------------------------------------------------------------------------ r72627 | vboxsync | 2018-06-20 21:53:28 +0800 (Wed, 20 Jun 2018) | 5 lines
Additions: relicence components needed for Linux shared folders to MIT.
bugref:9109: Shared folders: update to match in-kernel code more closely
This change makes the code on which the Linux kernel shared folder patch is
based MIT-licenced, so that the version in the Linux kernel can be too.  This
would make it easier to move code back and forth.
------------------------------------------------------------------------