commit_72682 ------------------------------------------------------------------------ r72682 | vboxsync | 2018-06-26 00:07:08 +0800 (Tue, 26 Jun 2018) | 5 lines
Installer/linux: avoid duplicate kernel module building on failure.
No bugref.
When we rebuilt kernel modules in vboxdrv.sh we called setup && start.  However
start already calls setup if the modules fail to load, so that in that case
setup was called a second time.  This change removes the first redundant call.
------------------------------------------------------------------------