commit_72889 ------------------------------------------------------------------------ r72889 | vboxsync | 2018-07-05 00:09:15 +0800 (Thu, 05 Jul 2018) | 5 lines
r0drv/linux: adjust kernel header file for an openSUSE 15.0 kernel back-port.
Linux 4.13.0 changed wait_queue_t to wait_queue_entry_t.  openSUSE 15.0 back-
ported this change to their 4.12.14 kernel.  Fortunately there was never an
official 4.12.14 (4.12.9 was the last), so we test for that instead of 4.13.0
and hope that no one else did this.
------------------------------------------------------------------------