commit_72946 ------------------------------------------------------------------------ r72946 | vboxsync | 2018-07-07 23:30:57 +0800 (Sat, 07 Jul 2018) | 1 line
Runtime/RTFuzz: Some updates, add a mode where the client is aware of being fuzzed for improved efficiency. The input data is fuzzed in the client and fed to the consumer until the program crashes upon the master can reconstruct the input causing the crash because we work with deterministic random number generators. This eliminates the overhead of constantly spawning new client processes. [scm]
------------------------------------------------------------------------