commit_73266 ------------------------------------------------------------------------ r73266 | vboxsync | 2018-07-20 22:27:20 +0800 (Fri, 20 Jul 2018) | 1 line
PGM,HM: Made PGMR3ChangeMode work in ring-0 too.  This required a kludge for the VT-x real-in-V86-mode stuff, as there are certain limitations on that mode which weren't checked as CR0.PE was cleared.  The kludge isn't very smart, but it seems to do the job.  Similar kludge for getting out of the mode. bugref:9044
------------------------------------------------------------------------