commit_73467 ------------------------------------------------------------------------ r73467 | vboxsync | 2018-08-03 17:49:55 +0800 (Fri, 03 Aug 2018) | 1 line
DrvAudio: Cleaned up the internal stream organization by not duplicating stuff like the status and the like by having two PDMAUDIOSTREAMs (on for the guest side and one for the host side). Instead, only use PDMAUDIOSTREAM and keep the guest and host specifics in PDMAUDIOSTREAMCTX. Should help simplifying things a lot.
------------------------------------------------------------------------