commit_73569 ------------------------------------------------------------------------ r73569 | vboxsync | 2018-08-08 23:20:23 +0800 (Wed, 08 Aug 2018) | 1 line
Audio/DrvAudio: Use capturing mixing buffers the same way as the playback ones: The child buffer (host side) is being used as intermediate buffer for conversion, whereas the parent buffer (guest side) will act as the actual ring buffer for the device emulation. For playback it's the other  way around.
------------------------------------------------------------------------