commit_73712 ------------------------------------------------------------------------ r73712 | vboxsync | 2018-08-16 21:32:54 +0800 (Thu, 16 Aug 2018) | 1 line
Devices/Serial: Remove giving the amount of available data to write from the UART core and let the drivers figure it out themselves when querying the data to write. Avoids unnecessary trips to R3 when the FIFO is enabled if the guest keeps the FIFO filled. Some other smaller fixes
------------------------------------------------------------------------