commit_73720 ------------------------------------------------------------------------ r73720 | vboxsync | 2018-08-17 01:57:44 +0800 (Fri, 17 Aug 2018) | 1 line
Runtime/serialport-win.cpp: Fixes, can't use CancelIo because it applies to the whole device and not only to the overlapped read (the required CancelIoEx is only available with Vista+). Luckily COMSTAT provides us with the necessary information to avoid CancelIo. Also set proper COM timeouts so we can get true non blocking behavior
------------------------------------------------------------------------