commit_73743 ------------------------------------------------------------------------ r73743 | vboxsync | 2018-08-18 01:56:34 +0800 (Sat, 18 Aug 2018) | 2 lines
Main/Progress+Appliance+Machine: Turn IProgress::waitForAsyncProgressCompletion into an internal method. It is not needed by any client code outside the API, it's simply is too special. Also include the error propagation which it originally skipped (leding to reduntant code in the calling code). Remove a replicated version of it from the Appliance code which seems to be the half-forgotten ancestor of the method in IProgress. Adapt the users of the method to the new internal method, which made the code easier to read.

------------------------------------------------------------------------