commit_74172 ------------------------------------------------------------------------ r74172 | vboxsync | 2018-09-10 16:17:28 +0800 (Mon, 10 Sep 2018) | 15 lines
Additions/drm: do not unpin an old frame buffer before the new one is in place.
bugref:9240: Additions/linux/vboxvideo: GPU memory management bug
We have seen sporadic cases of [unpin bad] messages in guest dmesg, and were
able to connect some of them to unpinning an old frame buffer when pinning a
new one before the new pin had succeeded.  If the new frame buffer was too
big to be pinned in the available video memory the pin operation failed, but
the old frame buffer was already unpinned.  As this did not get tracked in
the error path a subsequent unpin operation on it failed with the message
above.
This change fixes that by only unpinning the old frame buffer after the new
one has been successfully pinned.  While this potentially increases memory
pressure further, it avoids the case where we unpin the old one, fail to pin
the new one and fail to re-pin the old one, as well as the case of the old one
getting re-pinned at a different address.
Thank you Hans de Goede for review and comments/suggestions.
------------------------------------------------------------------------