commit_74259 ------------------------------------------------------------------------ r74259 | vboxsync | 2018-09-14 15:59:16 +0800 (Fri, 14 Sep 2018) | 7 lines
Additions/Solaris: adjust installer now that vboxvideo is upstream.
ticketref:17966: GuestAddition error
As of X.Org Server 1.19, Solaris provides a driver for the VirtualBox guest
graphics device itself, so we do not need to install it as part of the
Additions.  However the message in the installer did not reflect this.
This change corrects that and stops the installer returning failure in this
case as well.
------------------------------------------------------------------------