commit_74374 ------------------------------------------------------------------------ r74374 | vboxsync | 2018-09-20 00:08:07 +0800 (Thu, 20 Sep 2018) | 6 lines
Additions/linux: extend export modules to export without archiving.
bugref: 9165: webtools: create scripts for (re-)creating chroot build jails
This change adds a --folder <folder> syntax to export_modules.sh to tell it
to put the Additions modules build tree at a file system location instead of
in a tar archive.  This is for use with building in a number of chroots at one
time.
------------------------------------------------------------------------