commit_74412 ------------------------------------------------------------------------ r74412 | vboxsync | 2018-09-22 00:47:56 +0800 (Sat, 22 Sep 2018) | 8 lines
Additions/linux: remove attempt to suppress a log message.
The build logs for kernel modules start with a test for autoconf.h which
includes the text of an error message in the text.  This historically confused
users who thought it was a real error, so we tried to suppress it from our
logs, but the egrep command failed on Ubuntu 14.04.  Since though we now only
include the build log in case of error anyway there is no longer a risk that
users will think that a successful build failed, so just remove the
suppression.
------------------------------------------------------------------------