commit_74581 ------------------------------------------------------------------------ r74581 | vboxsync | 2018-10-02 22:01:09 +0800 (Tue, 02 Oct 2018) | 19 lines
Additions/linux/vboxvideo: better work-around for missing dirty rectangles.
bugref:9235: No screenshot for Linux VM
Background: a long time ago, Linux user-space applications using the DRM
graphics interface did not need to provide the graphics driver with dirty
rectangle information.  This was changed to make life easier for virtual
devices like ours, but the version of Plymouth (the splash screen at system
boot) in recent Ubuntu 14.04 versions does not do it.  We did some nasty
tricks to make the hardware notice screen updates anyway, and they caused
screenshots in Linux virtual machines not to work (and might have other side
effects).
This changes removes the tricks mentioned above, and instead introduces a
twice-a-second refresh timer in the driver.  As this is only intended for that
version of Plymouth, we disable the timer as soon as we either receive dirty
rectangle information, or the user-space client shows that it can handle
screen resizes (which Plymouth cannot, at least that version).  Once page flip
support is integrated we could use that as a test instead of the second, but
only if there is actually a need.  Twice a second refresh is quite a high
overhead, but it is only incurred in that rare case, and usually only for a
few seconds at that.
------------------------------------------------------------------------