commit_74622 ------------------------------------------------------------------------ r74622 | vboxsync | 2018-10-05 14:25:19 +0800 (Fri, 05 Oct 2018) | 6 lines
Graphics/VBoxVideo.h: make official that calling VBVA enable updates the screen.
bugref:9235: No screenshot for Linux VM
In the Linux Additions we make use of the fact that calling VBVA enable when
VBVA is already enabled causes the screen to update completely.  This is an
implementation detail, but it is easy to do on the host side, so make it
official by way of a comment in the right header file.
------------------------------------------------------------------------