commit_74779 ------------------------------------------------------------------------ r74779 | vboxsync | 2018-10-12 16:05:16 +0800 (Fri, 12 Oct 2018) | 14 lines
Additions/linux/vboxvideo: only set the views once, at initialisation time.
bugref: 9253: Enabling 2nd virtual screen results in an assert.
Views in vboxvideo/HGSMI are a concept used to tell the graphics card about
the virtual frame-buffer in use for each virtual screen.  Up until now we have
set these prior to each mode-set operation.  The assertion reported in this
bug was due to a mismatch when enabling and disabling screens using a mode-set
for DPMS.
In fact though, there is no good reason for the graphics hardware to need to
know about the frame-buffers at all.  It is enough for it to know the offset
in the video memory where scan-out starts, along with size and stride.  So
this change switches to just setting the views once, at driver initialisation
time, and setting them to cover the whole range of potentially usable video
memory for each screen at that point.  Thus whatever framebuffer is created
for a mode-set will always be a subset of the view.
------------------------------------------------------------------------