commit_74886 ------------------------------------------------------------------------ r74886 | vboxsync | 2018-10-17 23:21:01 +0800 (Wed, 17 Oct 2018) | 2 lines
Main/Appliance: mark the OCI specific code for moving to the extpack (it can't work to have specific code in the generic API, the rules will be always too volatile), and additionally fix lock order issues originating in the "too smart" code for figuring out the first disk. The previous code was getting the base medium which was kind of a workaround for limitations in the image upload logic, which were temporary and are about to be resolved.

------------------------------------------------------------------------